refactor: mover os testes de integração para respectivo diretório
