+++ package.json
@@ -34,7 +34,7 @@
     "dev": "cross-env NODE_ENV=serverest-development nodemon ./src/server.js --nodoc",
     "test:mutation": "cross-env NODE_ENV=serverest-test stryker run ./test/stryker.conf.js",
     "test:mutation:diff": "cross-env NODE_ENV=serverest-test stryker-diff-runner --path ./test/stryker.conf.js --branch trunk",
-    "test": "cross-env NODE_ENV=serverest-test nyc mocha --config test/.mocharc.js",
+    "test": "cross-env NODE_ENV=serverest-test nyc mocha --config test/integration/.mocharc.js",
     "lint": "standard",
     "lint:fix": "standard --fix",
     "lint:staged": "lint-staged",
