+++ test/integration/.mocharc.js
@@ -1,5 +1,5 @@
 module.exports = {
-  require: 'test/utils/global.js',
+  require: 'test/integration/utils/global.js',
   spec: ['test/**/*.test.js'],
   timeout: 30000
 }
