+++ test/integration/utils/global.js
@@ -1,4 +1,4 @@
-const app = require('../../src/app')
+const app = require('../../../src/app')
 const supertest = require('supertest')
 
 global.request = supertest(app)
